+++ lib/libesp32/berry_matter/src/be_matter_module.c
@@ -214,13 +214,13 @@ extern const bclass be_class_Matter_TLV;   // need to declare it upfront because
 #include "solidify/solidified_Matter_Plugin_1_Device.h"
 #include "solidify/solidified_Matter_Plugin_2_OnOff.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_OnOff.h"
-#include "solidify/solidified_Matter_Plugin_2_Light0.h"
+#include "solidify/solidified_Matter_Plugin_3_Light0.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Light0.h"
-#include "solidify/solidified_Matter_Plugin_3_Light1.h"
+#include "solidify/solidified_Matter_Plugin_2_Light1.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Light1.h"
-#include "solidify/solidified_Matter_Plugin_4_Light2.h"
+#include "solidify/solidified_Matter_Plugin_3_Light2.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Light2.h"
-#include "solidify/solidified_Matter_Plugin_4_Light3.h"
+#include "solidify/solidified_Matter_Plugin_3_Light3.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Light3.h"
 #include "solidify/solidified_Matter_Plugin_2_Shutter.h"
 #include "solidify/solidified_Matter_Plugin_3_ShutterTilt.h"
